Fixes #38830 - Add sub-menu Host collections to bulk hosts Change associations
